+++ consensus/clique/src/test/java/org/hyperledger/besu/consensus/clique/CliqueProtocolScheduleTest.java
@@ -22,6 +22,9 @@ import static org.mockito.Mockito.when;
 import org.hyperledger.besu.config.CliqueConfigOptions;
 import org.hyperledger.besu.config.GenesisConfigFile;
 import org.hyperledger.besu.config.GenesisConfigOptions;
+import org.hyperledger.besu.config.JsonCliqueConfigOptions;
+import org.hyperledger.besu.consensus.common.ForkSpec;
+import org.hyperledger.besu.consensus.common.ForksSchedule;
 import org.hyperledger.besu.cryptoservices.NodeKey;
 import org.hyperledger.besu.cryptoservices.NodeKeyUtils;
 import org.hyperledger.besu.datatypes.Hash;
@@ -35,6 +38,7 @@ import org.hyperledger.besu.ethereum.mainnet.ProtocolSpec;
 import org.hyperledger.besu.evm.internal.EvmConfiguration;
 
 import java.time.Instant;
+import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
@@ -57,7 +61,12 @@ public class CliqueProtocolScheduleTest {
     final GenesisConfigOptions config = GenesisConfigFile.fromConfig(jsonInput).getConfigOptions();
     final ProtocolSchedule protocolSchedule =
         CliqueProtocolSchedule.create(
-            config, NODE_KEY, false, EvmConfiguration.DEFAULT, new BadBlockManager());
+            config,
+            new ForksSchedule<>(List.of()),
+            NODE_KEY,
+            false,
+            EvmConfiguration.DEFAULT,
+            new BadBlockManager());
 
     final ProtocolSpec homesteadSpec = protocolSchedule.getByBlockHeader(blockHeader(1));
     final ProtocolSpec tangerineWhistleSpec = protocolSchedule.getByBlockHeader(blockHeader(2));
@@ -71,9 +80,12 @@ public class CliqueProtocolScheduleTest {
 
   @Test
   public void parametersAlignWithMainnetWithAdjustments() {
+    final ForksSchedule<CliqueConfigOptions> forksSchedule =
+        new ForksSchedule<>(List.of(new ForkSpec<>(0, JsonCliqueConfigOptions.DEFAULT)));
     final ProtocolSpec homestead =
         CliqueProtocolSchedule.create(
                 GenesisConfigFile.DEFAULT.getConfigOptions(),
+                forksSchedule,
                 NODE_KEY,
                 false,
                 EvmConfiguration.DEFAULT,
@@ -88,7 +100,7 @@ public class CliqueProtocolScheduleTest {
 
   @Test
   public void zeroEpochLengthThrowsException() {
-    final CliqueConfigOptions cliqueOptions = mock(CliqueConfigOptions.class);
+    final CliqueConfigOptions cliqueOptions = mock(JsonCliqueConfigOptions.class);
     when(cliqueOptions.getEpochLength()).thenReturn(0L);
     when(genesisConfig.getCliqueConfigOptions()).thenReturn(cliqueOptions);
 
@@ -96,6 +108,7 @@ public class CliqueProtocolScheduleTest {
             () ->
                 CliqueProtocolSchedule.create(
                     genesisConfig,
+                    new ForksSchedule<>(List.of()),
                     NODE_KEY,
                     false,
                     EvmConfiguration.DEFAULT,
@@ -106,7 +119,7 @@ public class CliqueProtocolScheduleTest {
 
   @Test
   public void negativeEpochLengthThrowsException() {
-    final CliqueConfigOptions cliqueOptions = mock(CliqueConfigOptions.class);
+    final CliqueConfigOptions cliqueOptions = mock(JsonCliqueConfigOptions.class);
     when(cliqueOptions.getEpochLength()).thenReturn(-3000L);
     when(genesisConfig.getCliqueConfigOptions()).thenReturn(cliqueOptions);
 
@@ -114,6 +127,7 @@ public class CliqueProtocolScheduleTest {
             () ->
                 CliqueProtocolSchedule.create(
                     genesisConfig,
+                    new ForksSchedule<>(List.of()),
                     NODE_KEY,
                     false,
                     EvmConfiguration.DEFAULT,
@@ -131,9 +145,16 @@ public class CliqueProtocolScheduleTest {
         "{\"config\": " + "\t{\"chainId\": 1337,\n" + "\t\"londonBlock\": 2}\n" + "}";
 
     final GenesisConfigOptions config = GenesisConfigFile.fromConfig(jsonInput).getConfigOptions();
+    final ForksSchedule<CliqueConfigOptions> forksSchedule =
+        new ForksSchedule<>(List.of(new ForkSpec<>(0, JsonCliqueConfigOptions.DEFAULT)));
     final ProtocolSchedule protocolSchedule =
         CliqueProtocolSchedule.create(
-            config, NODE_KEY, false, EvmConfiguration.DEFAULT, new BadBlockManager());
+            config,
+            forksSchedule,
+            NODE_KEY,
+            false,
+            EvmConfiguration.DEFAULT,
+            new BadBlockManager());
 
     BlockHeader emptyFrontierParent =
         headerBuilder